naming-figma-layer-engine/requirements.txt
pytorch
pytorch-lightning
transformers